+++ cmd/template/app/runner.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 	"text/template"
 
+	"github.com/giantswarm/k8smetadata/pkg/label"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/spf13/cobra"
@@ -73,6 +74,7 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		CatalogNamespace:  r.flag.CatalogNamespace,
 		Cluster:           clusterName,
 		DefaultingEnabled: r.flag.DefaultingEnabled,
+		ExtraLabels:       map[string]string{},
 		InCluster:         r.flag.InCluster,
 		Name:              r.flag.Name,
 		Namespace:         targetNamespace,
@@ -110,6 +112,13 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		}
 		appConfig.UserConfigSecretName = userSecret.GetName()
 
+		if r.flag.PreventDeletion {
+			if userSecret.Labels == nil {
+				userSecret.Labels = map[string]string{}
+			}
+			userSecret.Labels[label.PreventDeletion] = "true" //nolint:goconst
+		}
+
 		userConfigSecretYaml, err = yaml.Marshal(userSecret)
 		if err != nil {
 			return microerror.Mask(err)
@@ -129,6 +138,13 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		}
 		appConfig.UserConfigConfigMapName = userConfigMap.GetName()
 
+		if r.flag.PreventDeletion {
+			if userConfigMap.Labels == nil {
+				userConfigMap.Labels = map[string]string{}
+			}
+			userConfigMap.Labels[label.PreventDeletion] = "true"
+		}
+
 		userConfigConfigMapYaml, err = yaml.Marshal(userConfigMap)
 		if err != nil {
 			return microerror.Mask(err)
@@ -146,6 +162,10 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		return microerror.Mask(err)
 	}
 	appConfig.NamespaceConfigLabels = namespaceLabels
+
+	if r.flag.PreventDeletion {
+		appConfig.ExtraLabels[label.PreventDeletion] = "true"
+	}
 
 	r.setTimeouts(&appConfig)