+++ cmd/update/app/runner_test.go
@@ -98,6 +98,7 @@ func Test_run(t *testing.T) {
 	}
 
 	for i, tc := range testCases {
+		tc := tc
 		t.Run(fmt.Sprintf("case %d: %s", i, tc.name), func(t *testing.T) {
 			var err error